Build sequence diagram from trace
This commit creates a new endpoint `GET /api/traces/:trace-id/sequence-diagram` that fetches a trace from Jaeger and builds a sequence diagram from it. This implementation is based on the following assumptions:
- Every span has a tag `span.kind` that allow us to identify if it is a client or a server (more info at https://opentracing.io/specification/conventions/);
- Each trace process should become a lifeline;
- Server spans should become execution boxes on the sequence diagram;
- Client spans should become two arrows (out and in) in the sequence diagram';

This is a very basic implementation and only supports two kinds of span - `client` and `server`. In a future we should add support for others kinds, like `consumer` and `producer`.

Co-authored-by: Davi Correia Jr <davicorreiajr@gmail.com>
Co-authored-by: Pedro Darvas <pedro.darvas@usp.br>