Fork cljfmt linter so we get vertical alignment on maps and bindings
As of now cljfmt linter lacks vertical alignment on maps and bindings as discussed in https://github.com/weavejester/cljfmt/pull/77. So I decided to fork the original project and implement this feature. While this feature is not merged into the original project, I'm switching Atlas linter to this fork https://github.com/lucasmafra/cljfmt/.